requirements.txt
# Core Dependencies
speechrecognition==3.10.0
pyaudio==0.2.14
sounddevice==0.4.6
vosk==0.3.45
google-cloud-speech==2.21.0
pyttsx3==2.90

# AI and NLP
nltk==3.8.1
transformers==4.33.2
torch==2.0.1
numpy==1.24.3
scipy==1.11.2

# Web Framework and API
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
requests==2.31.0
httpx==0.24.1

# Database and Storage
sqlalchemy==2.0.20
alembic==1.12.0

# Image and Media Processing
Pillow==10.0.0
pygame==2.5.2
opencv-python==4.8.0.76

# System Integration
psutil==5.9.5
pywin32==306
plyer==2.1.0
watchdog==3.0.0

# Security and Authentication
PyJWT==2.8.0
cryptography==41.0.4
keyring==24.2.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3

# Email and Communication
email-validator==2.0.0
imaplib2==3.6

# Weather and External APIs
requests-oauthlib==1.3.1
python-weather==2.0.7
pyowm==3.3.0

# File Operations
send2trash==1.8.2
pathlibutil==0.2.0

# Logging and Monitoring
loguru==0.7.2
rich==13.5.2

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Build and Packaging
pyinstaller==5.13.2
nuitka==1.7.10

# Development Tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.4.0

# Wikipedia API
wikipedia==1.4.0

# Utilities
click==8.1.7
rich==13.5.2
typer==0.9.0
pytz==2023.3
python-dateutil==2.8.2